Changes for testing doze and standby modes
Added the possibility to set the peer name in settings. When the peer name is empty, the library will use the simple handshake message whose length is only one byte. The received handshake message can be either the short, simple or the long one - both are validated the same. This is essential when testing together with the Thali Cordova Plug-in unit test app.